separate SAPT from psikit.py
