+++ psikit/__init__.py
@@ -1,3 +1,2 @@
 # -*- coding: utf-8 -*-
 from psikit.psikit import Psikit
-from psikit.psikit import Sapt
