+++ tests/test_init.py
@@ -2,7 +2,7 @@ import sys
 import os
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
 from psikit import Psikit
-from psikit import Sapt
+from psikit import sapt
 import rdkit
 import pytest
